% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CLASSIFICATION}
\alias{CLASSIFICATION}
\title{CLASSIFICATION Table}
\format{
A data frame with 841 rows and 3 variables:
\describe{
\item{\code{scrape_datetime}}{character COLUMN_DESCRIPTION}
\item{\code{concept_classification}}{character COLUMN_DESCRIPTION}
\item{\code{rn_url}}{character COLUMN_DESCRIPTION}
}
}
\usage{
CLASSIFICATION
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
