% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONCEPT_ANCESTOR}
\alias{CONCEPT_ANCESTOR}
\title{CONCEPT_ANCESTOR}
\format{
A data frame with 16980 rows and 4 variables:
\describe{
\item{\code{ancestor_concept_id}}{character COLUMN_DESCRIPTION}
\item{\code{descendant_concept_id}}{character COLUMN_DESCRIPTION}
\item{\code{min_levels_of_separation}}{character COLUMN_DESCRIPTION}
\item{\code{max_levels_of_separation}}{character COLUMN_DESCRIPTION}
}
}
\usage{
CONCEPT_ANCESTOR
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
