% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONCEPT_RELATIONSHIP}
\alias{CONCEPT_RELATIONSHIP}
\title{CONCEPT_RELATIONSHIP}
\format{
A data frame with 47 rows and 6 variables:
\describe{
\item{\code{concept_id_1}}{integer COLUMN_DESCRIPTION}
\item{\code{concept_id_2}}{integer COLUMN_DESCRIPTION}
\item{\code{relationship_id}}{character COLUMN_DESCRIPTION}
\item{\code{valid_start_date}}{double COLUMN_DESCRIPTION}
\item{\code{valid_end_date}}{double COLUMN_DESCRIPTION}
\item{\code{invalid_reason}}{character COLUMN_DESCRIPTION}
}
}
\usage{
CONCEPT_RELATIONSHIP
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
