% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OLD_RN_URL_VALIDITY}
\alias{OLD_RN_URL_VALIDITY}
\title{OLD_RN_URL_VALIDITY}
\format{
A data frame with 2328 rows and 3 variables:
\describe{
\item{\code{rnuv_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{rn_url}}{character COLUMN_DESCRIPTION}
\item{\code{is_404}}{character COLUMN_DESCRIPTION}
}
}
\usage{
OLD_RN_URL_VALIDITY
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
