% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHRASE_LOG}
\alias{PHRASE_LOG}
\title{PHRASE_LOG Table}
\format{
A data frame with 3024 rows and 11 variables:
\describe{
\item{\code{phrase_datetime}}{character COLUMN_DESCRIPTION}
\item{\code{input}}{character COLUMN_DESCRIPTION}
\item{\code{phrase}}{character COLUMN_DESCRIPTION}
\item{\code{type}}{character COLUMN_DESCRIPTION}
\item{\code{url}}{character COLUMN_DESCRIPTION}
\item{\code{response_received}}{character COLUMN_DESCRIPTION}
\item{\code{no_record}}{character COLUMN_DESCRIPTION}
\item{\code{response_recorded}}{character COLUMN_DESCRIPTION}
\item{\code{compound_match}}{character COLUMN_DESCRIPTION}
\item{\code{rn}}{character COLUMN_DESCRIPTION}
\item{\code{rn_url}}{character COLUMN_DESCRIPTION}
}
}
\usage{
PHRASE_LOG
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
