% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SYNONYMS}
\alias{SYNONYMS}
\title{SYNONYMS Table}
\format{
A data frame with 5840 rows and 4 variables:
\describe{
\item{\code{scrape_datetime}}{character COLUMN_DESCRIPTION}
\item{\code{rn_url}}{character COLUMN_DESCRIPTION}
\item{\code{concept_synonym_type}}{character COLUMN_DESCRIPTION}
\item{\code{concept_synonym_name}}{character COLUMN_DESCRIPTION}
}
}
\usage{
SYNONYMS
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
