% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNMATCHED_INVEST_DRUGS5}
\alias{UNMATCHED_INVEST_DRUGS5}
\title{UNMATCHED_INVEST_DRUGS5}
\format{
A data frame with 4738 rows and 4 variables:
\describe{
\item{\code{id_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{invest_drug_rowid}}{double COLUMN_DESCRIPTION}
\item{\code{term_type}}{character COLUMN_DESCRIPTION}
\item{\code{terms}}{character COLUMN_DESCRIPTION}
}
}
\usage{
UNMATCHED_INVEST_DRUGS5
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
